N_HANS___Source_Separation/requirements.txt
numpy>=1.14.5
tensorflow_gpu>=1.14.0
scipy>=1.0.1
